+++ src/main/java/com/hackaton/Good_Night_3rd_Hackathon_Backend/wish/entity/Wish.java
@@ -13,7 +13,7 @@ import java.sql.Timestamp;
 @Setter
 @NoArgsConstructor
 public class Wish {
-    private int wishId;
+    private Long wishId;
     private String title;
     private String content;
     private String category;
